Add a simple query command
This patch adds a basic query command to help play with the
index before having the whole api up.

> Search for recipes, from the command line!
>
> USAGE:
>     cantine query <query> --base-dir <base_dir>
>
> FLAGS:
>     -h, --help       Prints help information
>     -V, --version    Prints version information
>
> OPTIONS:
>     -b, --base-dir <base_dir>    Path to the data built by `load`
>
> ARGS:
>     <query>    Plain text query to submit

This will evolve into something that takes a SearchQuery as
an input probably, but for now it was enough to show that I
can keep on with the idea of driving things separately.