Stabilize the QueryParser result
This patch changes the query parser in a way that it doesn't come
loaded with business logic ("how to handle an empty query?"), nor
yields weird stuff like 0/1-sized clauses BooleanQuery.

Now it's trivial to decide how to handle queries because we will
always get `Ok(None)` when the query is effectively empty. Before
this patch we'd get a `MatchAllDocsQuery` for empty queries, None
for single term queries excluded by the tokenizer and a
possibly-weirdly-sized boolean query.