Make the score generic, but only implement for f32
This patch renames StableComparableDoc to Scored and makes everything
that works with tantivy's Score (an alias to f32) accept a type
parameter for the score.

The only reason this looks so trivial is because there's still only
the existing implementation for f32. Making this generic over T
requires a lot more verbosity to satisfy the compiler.