Use a condition trait as the new SearchMarker
This patch builds on what I learned writing the `custom_collector`
mod to replace the `after` functionality with something more
generic.

This allows implementing arbitrary skipping logic (a user-provided
exclusion list, for example). I'll follow up with using the same
idea as in `DocScorerFactory` so that one can leverage the index
data when checking the conditions too.