Experiment holding the Config impl
That's a way nicer API, but it doesn't do what I want on `add`.

I lose control of `first` and `second` in the test: I can't wrap
it in an optional and test it against a database result, can't even
drop it without having the whole database.

So it looks like it's better to split the config in two and then
I can hold the decoder config and require the encoder as a type
parameter like in the first try.