Store floats in a u64 field
I still hit the fast field crash when indexing fast float fieds
in parallel. My guess is that somewhere along the code it assumes
as fast field readers are loaded (maybe prior to the second merging
thread?) but floats are relatively rare in model and then it blows
up.

Anyway, I'll investigate that later - let's get this working first.