Transform query into a "api" service
No functional changes, the only added code is for wiring actix.

The number of added dependencies is very scary. But other than that,
it's really easy to drive actix, so the work from now on should be
super easy. Nice.

Now it's a good time to tame these errors I think. Then I can finally
port the web site.