Add types back to cursors
This patch starts adding the type information to the cursors. This
information is"lost" by containing everything as (u64,u64), so we
add it back by making cursor types enums.

The price I'm paying is an added byte for the type tag (mapping to
an enum member) which leads to 2 extra characters in the base64
SearchCursor representation.