Add broad filters based on sort parameters
When sorting by a feature, only output recipes that actually contain
said feature.

I'm not sure I like this - In one hand the `items` part of the output
is a lot more sane, but it's really weird to ask for sorting and
having the result-set change drastically.

I think the ideal, albeit still very weird, is pushing the items
with missing features to the end of the results (i.e. score them
with MAX or MIN, based on order), but doing that externally will
be very annoying.