Use `Option<After>` instead of having `After::Start`
The main motivation of this is dropping the very awkward `is_start` /
`INVALID_RECIPE_ID` thingers. The result is some convoluted code that
I'm having a really hard time expressing as traits.

I think my problem is that in order to decide which collector to use
we need to look at 3 things: Sort, After and Order. It looks like if
I flatten order into the sort enum (so, doubling the sort enum size),
things will at least read more easily, so I'll try that soon.