Start working on a order-aware typed collector
Having to do the `Reverse` dance only to flip the scores again when
collecting in ascending order was bothering me more than it probably
should. This patch starts sketching what ultimately become something
like this:

    OrderedCollector::<Score, topk::Descending, _>::with_limit(10, true);

Meaning: you control the order via the type instead of creating a
custom scorer/tweaker that flips the score before collecting. Whilst
the motivation for this is purely cosmetic, it sounds like it should
be faster for very hot loops (far from my use-case of ~1M docs)

This is the 3rd time I'm attempting this and looks like I'll finally
be able to pull it off :tada: