Move order-aware functionality into a separate module
This patch moves the `OrderedCollector` (and family) into a new
module. Since I'll be going through all the code inside the
`top_collector` mod, I'm taking the opportunity to clean up
a bit the interfaces.

This has helped me in seeing a weakness in removing specialized
types from the CollectionResult: I can't simply submit a result
item as a sort condition for pagination since it's necessary
to know how to compare them.

I'll follow up trying to add types back in, but now I'm not as
confident that this is a very polish-able turd :-(