Make CheckCondition<T> aware of the order
This looks totally like a wart and I think I'd prefer a more explicit
provider-driven checker trait.

I'll try to further encode it in the type, but I think I won't be
able to make it easy to hold. Essentially, I'd need to move the condition
trait into the provider, which would force it to be a struct holding
_something_ (a `PhantomData` marker, likely) which might make it
difficult to infer the type.