Drive the search logic via the new `TopCollector`
This patch switches to the new `conditional_collector` module, which
has built-in support for changing the order. This means that we need
a lot less business logic for each possible state (sort, order and
pagination state) and it all can be expressed via `TopCollector`.

So much nicer now :-) With some polish, `tique::conditional_collector`
is a good candidate for public release.