Read options solely from `std::env`
StructOpt + clap are great and all, but I drive everything with
environment variables in production already and am more comfortable
with the language, so spending extra compilation time for the sake
of a cute CLI interface is not worth it anymore.

My 2013 CPU is thankful. Now my biggest gripe WRT build is that
`tantivy` pulls `failure` which implies a build dependency on GCC
for the sake of stacktraces... So xbuild for ARM is a bit annoying.