Stop generating custom collector code
It works! My only gripe is that this new API loses the ability to
use an embedded database (say, a hashmap) to read the metadata
since it requires static lifetimes everywhere.

I'm focusing on bytes fast fields, so requiring ownership is fine
for now. It would be nice to implement `Feature<Query> for &Input`
and use the same collector code tho, so we'll see...