Initial TopTerms implementation
TopTerms reads the index and extracts the most relevant terms in a
given document or any arbitrary text input. You can use it to build
keywords for your documents or, more interestingly, use the result
as a query to find similar documents.

It's pretty much a reimplementation of Lucene's MoreLikeThis. I don't
particularly like this approach in prod (too many knobs, dependency on
the index to formulate a query), but it yields pretty good results
with little effort.

Ref: https://lucene.apache.org/core/8_4_1/queries/org/apache/lucene/queries/mlt/MoreLikeThis.html