Use multiple fields when indexing and querying
This patch splits the `fulltext` index field into its sources: name,
ingredients and instructions, making them available for individual
strict querying.

Unsurprisingly, using an OR-by-default boolean-query-based with multiple
fields yields awful results for simple queries. To account for that, we
switch the query parser to use dismax with a 10% tiebreaking increment.